+++ packages/api/src/controllers/server.controller.ts
@@ -37,6 +37,7 @@ import {
   UpdateUserData,
 } from '../domains/user/transfer-objects/user.dto'
 import { transformUserToUserDTO } from '../domains/user/transforms/user-dto.transform'
+import { UnauthorizedError } from '../errors/auth.error'
 import type { ErrorResponse } from '../transfer-objects/error-response.dto'
 
 export interface ListUsersResponse {
@@ -63,9 +64,12 @@ export class ServerController extends Controller {
     @Request() req: Express.Request,
     @Path() locationType: ServerLocationType,
   ): Promise<ServerLocationData[]> {
+    if (!req.user) {
+      throw new UnauthorizedError()
+    }
     const results =
       await this.serverConfigurationService.listConfiguredServerLocationsAsUser(
-        req.viewer.id,
+        req.user.id,
         locationType,
       )
 
@@ -80,9 +84,12 @@ export class ServerController extends Controller {
     @Body() payload: ServerLocationInputData,
     @Path() locationType: ServerLocationType,
   ): Promise<ServerLocationData> {
+    if (!req.user) {
+      throw new UnauthorizedError()
+    }
     const record =
       await this.serverConfigurationService.addServerLocationServerConfigurationAsUser(
-        req.viewer.id,
+        req.user.id,
         locationType,
         payload,
       )
@@ -106,10 +113,12 @@ export class ServerController extends Controller {
     @Path() locationType: ServerLocationType,
     @Path() locationId: string,
   ) {
-    const _user = await this.userService.getById({ id: req.viewer.id })
+    if (!req.user) {
+      throw new UnauthorizedError()
+    }
 
     await this.serverConfigurationService.deleteServerLocationServerConfigurationAsUser(
-      req.viewer.id,
+      req.user.id,
       locationType,
       locationId,
     )
@@ -121,8 +130,11 @@ export class ServerController extends Controller {
   @OperationId('listUsers')
   @Get('/users')
   async listUsers(@Request() req: Express.Request): Promise<ListUsersResponse> {
+    if (!req.user) {
+      throw new UnauthorizedError()
+    }
     const { results, totalCount } = await this.userService.listUsersAsAdmin(
-      req.viewer.id,
+      req.user.id,
       {
         limit: 100,
         offset: 0,
@@ -141,8 +153,11 @@ export class ServerController extends Controller {
     @Request() req: Express.Request,
     @Path() userId: string,
   ): Promise<{ result: UserData }> {
+    if (!req.user) {
+      throw new UnauthorizedError()
+    }
     const result = await this.userService.getUserByIdAsAdmin(
-      req.viewer.id,
+      req.user.id,
       userId,
     )
     return {
@@ -158,8 +173,11 @@ export class ServerController extends Controller {
     @Body()
     body: CreateUserData,
   ) {
-    const user = await this.userService.createUserAsAdmin(req.viewer, body)
-    return { user: transformUserToUserDTO(user) }
+    if (!req.user) {
+      throw new UnauthorizedError()
+    }
+    const createdUser = await this.userService.createUserAsAdmin(req.user, body)
+    return { user: transformUserToUserDTO(createdUser) }
   }
 
   @Security(AuthScheme.AccessToken, [AuthScope.CreateUsers])
@@ -171,18 +189,26 @@ export class ServerController extends Controller {
     @Body()
     body: UpdateUserData,
   ) {
-    const user = await this.userService.updateUserAsAdmin(req.viewer, {
+    if (!req.user) {
+      throw new UnauthorizedError()
+    }
+
+    const updatedUser = await this.userService.updateUserAsAdmin(req.user, {
       id: userId,
       ...body,
     })
-    return { user: transformUserToUserDTO(user) }
+    return { user: transformUserToUserDTO(updatedUser) }
   }
 
   @Security(AuthScheme.AccessToken, [AuthScope.CreateUsers])
   @OperationId('deleteUser')
   @Delete('/users/:userId')
   async deleteUser(@Request() req: Express.Request, @Path() userId: string) {
-    await this.userService.deleteUserAsAdmin(req.viewer, userId)
+    if (!req.user) {
+      throw new UnauthorizedError()
+    }
+
+    await this.userService.deleteUserAsAdmin(req.user, userId)
     return true
   }
 
@@ -192,9 +218,12 @@ export class ServerController extends Controller {
   async getSettings(
     @Request() req: Express.Request,
   ): Promise<{ settings: ServerSettings }> {
+    if (!req.user) {
+      throw new UnauthorizedError()
+    }
     return {
       settings: await this.serverConfigurationService.getServerSettingsAsUser(
-        req.viewer,
+        req.user,
       ),
     }
   }
@@ -207,14 +236,18 @@ export class ServerController extends Controller {
     @Path() settingsKey: string,
     @Body() settingsValue: { value: any },
   ): Promise<{ settings: ServerSettings }> {
+    if (!req.user) {
+      throw new UnauthorizedError()
+    }
+
     await this.serverConfigurationService.setServerSettingAsUser(
-      req.viewer,
+      req.user,
       settingsKey,
       settingsValue.value,
     )
     return {
       settings: await this.serverConfigurationService.getServerSettingsAsUser(
-        req.viewer,
+        req.user,
       ),
     }
   }
@@ -226,13 +259,16 @@ export class ServerController extends Controller {
     @Request() req: Express.Request,
     @Path() settingsKey: string,
   ): Promise<{ settings: ServerSettings }> {
+    if (!req.user) {
+      throw new UnauthorizedError()
+    }
     await this.serverConfigurationService.resetServerSettingAsUser(
-      req.viewer,
+      req.user,
       settingsKey,
     )
     return {
       settings: await this.serverConfigurationService.getServerSettingsAsUser(
-        req.viewer,
+        req.user,
       ),
     }
   }
@@ -242,8 +278,11 @@ export class ServerController extends Controller {
   @OperationId('createServerWorkerKey')
   @Post('/worker-keys')
   async createServerWorkerKey(@Request() req: Express.Request) {
+    if (!req.user) {
+      throw new UnauthorizedError()
+    }
     const result = await this.serverWorkerService.createServerWorkerKeyAsAdmin(
-      req.viewer,
+      req.user,
     )
     return {
       token: result.token,
@@ -259,8 +298,11 @@ export class ServerController extends Controller {
     @Request() req: Express.Request,
     @Path() workerKeyId: string,
   ) {
+    if (!req.user) {
+      throw new UnauthorizedError()
+    }
     await this.serverWorkerService.deleteServerWorkerKeyAsAdmin(
-      req.viewer,
+      req.user,
       workerKeyId,
     )
     return {
@@ -278,8 +320,11 @@ export class ServerController extends Controller {
     @Query() limit?: number,
     @Query() offset?: number,
   ) {
+    if (!req.user) {
+      throw new UnauthorizedError()
+    }
     const result = await this.serverWorkerService.listServerWorkerKeysAsAdmin(
-      req.viewer,
+      req.user,
       {
         limit,
         offset,
@@ -304,8 +349,11 @@ export class ServerController extends Controller {
     @Query() limit?: number,
     @Query() offset?: number,
   ) {
+    if (!req.user) {
+      throw new UnauthorizedError()
+    }
     const result = await this.serverWorkerService.listServerWorkersAsAdmin(
-      req.viewer,
+      req.user,
       {
         limit,
         offset,
